backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# LEGALBOT - DEPENDENCIAS DEL BACKEND
# ═══════════════════════════════════════════════════════════════

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # Para PostgreSQL
aiosqlite==0.19.0  # Para SQLite (demo sin servidor)
alembic==1.13.1

# ═══════════════════════════════════════════════════════════════
# AI and LLM (REQUERIDO PARA IA REAL)
# ═══════════════════════════════════════════════════════════════
openai==1.10.0

# Vector Database (REQUERIDO PARA RAG)
pinecone>=3.0.0

# Opcional: LangChain para flujos más complejos
# langchain==0.1.4
# langchain-openai==0.0.4

# ═══════════════════════════════════════════════════════════════
# Document Generation
# ═══════════════════════════════════════════════════════════════
python-docx==1.1.0
reportlab==4.0.8
jinja2==3.1.3

# Redis and Caching (Opcional)
redis==5.0.1

# HTTP Client (REQUERIDO para proveedores alternativos)
httpx==0.26.0
aiohttp==3.9.1

# Embeddings locales (Rápido y ligero, reemplaza a sentence-transformers)
fastembed==0.2.0

# Validation and Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Utils
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Payments (Para Culqi - Perú)
# culqi-python==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# PDF Processing
pymupdf>=1.23.0

